feat(front): 経歴入力欄をコンポーネント化
